Split out LibeventThread and friends into a separate header file
