+++ src/include/wire/libevent_server.h
@@ -12,27 +12,28 @@
 
 #pragma once
 
-#include <event2/listener.h>
-#include <event2/bufferevent.h>
 #include <event2/buffer.h>
+#include <event2/bufferevent.h>
 #include <event2/event.h>
+#include <event2/listener.h>
 
 #include <arpa/inet.h>
 #include <netinet/tcp.h>
 
 #include <signal.h>
-#include <string.h>
-#include <stdlib.h>
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 #include <iostream>
 #include <vector>
 
-#include "common/logger.h"
-#include "common/config.h"
-#include "common/exception.h"
-#include "container/lock_free_queue.h"
 #include <sys/file.h>
 #include <fstream>
+#include "common/config.h"
+#include "common/exception.h"
+#include "common/logger.h"
+#include "container/lock_free_queue.h"
+#include "wire/libevent_thread.h"
 #include "wire/wire.h"
 
 #define QUEUE_SIZE 100
@@ -40,6 +41,9 @@
 
 namespace peloton {
 namespace wire {
+
+// Forward Declarations
+class LibeventThread;
 
 // Libevent Thread States
 enum ConnState {
@@ -119,22 +123,14 @@ struct Buffer {
   }
 
   // single buffer element accessor
-  inline uchar GetByte(size_t &index) {
-    return buf[index];
-  }
+  inline uchar GetByte(size_t &index) { return buf[index]; }
 
   // Get pointer to index location
-  inline uchar* GetPtr(size_t index) {
-    return &buf[index];
-  }
+  inline uchar *GetPtr(size_t index) { return &buf[index]; }
 
-  inline ByteBuf::const_iterator Begin() {
-    return std::begin(buf);
-  }
+  inline ByteBuf::const_iterator Begin() { return std::begin(buf); }
 
-  inline ByteBuf::const_iterator End() {
-    return std::end(buf);
-  }
+  inline ByteBuf::const_iterator End() { return std::end(buf); }
 
   inline size_t GetMaxSize() { return SOCKET_BUFFER_SIZE; }
 };
@@ -149,65 +145,6 @@ struct NewConnQueueItem {
       : new_conn_fd(new_conn_fd),
         event_flags(event_flags),
         init_state(init_state) {}
-};
-
-class LibeventThread {
- protected:
-  // The connection thread id
-  const int thread_id_;
-  struct event_base *libevent_base_;
-
- public:
-  LibeventThread(const int thread_id, struct event_base *libevent_base)
-      : thread_id_(thread_id), libevent_base_(libevent_base) {
-    if (libevent_base_ == nullptr) {
-      LOG_ERROR("Can't allocate event base\n");
-      exit(1);
-    }
-  };
-
-  inline struct event_base *GetEventBase() { return libevent_base_; }
-
-  inline int GetThreadID() const { return thread_id_; }
-
-  // TODO implement destructor
-  inline ~LibeventThread() {}
-};
-
-class LibeventWorkerThread : public LibeventThread {
- private:
-  // New connection event
-  struct event *new_conn_event_;
-
- public:
-  // Notify new connection pipe(send end)
-  int new_conn_send_fd;
-
-  // Notify new connection pipe(receive end)
-  int new_conn_receive_fd;
-
-  /* The queue for new connection requests */
-  LockFreeQueue<std::shared_ptr<NewConnQueueItem>> new_conn_queue;
-
- public:
-  LibeventWorkerThread(const int thread_id);
-};
-
-class LibeventMasterThread : public LibeventThread {
- private:
-  const int num_threads_;
-
-  // TODO: have a smarter dispatch scheduler
-  int next_thread_id_ = 0; // next thread we dispatched to
-
- public:
-  LibeventMasterThread(const int num_threads, struct event_base *libevent_base);
-
-  void DispatchConnection(int new_conn_fd, short event_flags);
-
-  std::vector<std::shared_ptr<LibeventWorkerThread>> &GetWorkerThreads();
-
-  static void StartWorker(peloton::wire::LibeventWorkerThread *worker_thread);
 };
 
 /*
